requirements.txt
fastapi
uvicorn
pdfplumber
pandas
openai>=1.30.0
tenacity
python-dotenv